Add HTML, JS, CSS file. Create calculator basic functions.
Created basic calculator visual in HTML. In JS file added basic functionality
for a two number calculation. Created functions for the different operators,
like divide etc. Created an operate function that checks for the operator and
calls one of the operator functions on the numbers input. Added event listeners
for the buttons of the calculator. Created variables that stores the first and second
number input on the calculator and a variable that stores the operator clicked on
the calculator. Added an event listener for the equals button that takes the numbers
input and the operator and calls the operate function on it, to return the result
on the display div.

[Ticket: X]