Add more functionality to calculator, solve errors.
Added multiple functions: A decimal button, a backspace button, a clear button,
a function that limits calculations to 5 decimal points. Created a loop for the
number buttons, to eliminate duplicate code. Some error checking: Dividing by 0
which would create an error, now displays a warning message. Clicking the equal sign
when it was just used doesn't cause an error, it does nothing then. Clicking on
a number after clicking the equal sign doesn't add that number to the calculated
number on display anymore, it now resets the calculator and starts with the inputted
number.

[Ticket: X]