Fix errors, add styling to buttons
Fixed the error that happened when a number over a 1000 would enter a calculation.
It was in the roundToDecimal function where it put commas in numbers over a 1000,
which would then be used as the first number in a further calculation, such a number
wasn't able to be properly read by the operate function, which caused the error.
Added styling to buttons, so it feels a bit more interactive.

[Ticket: X]